Metrics: Add JSON serialization support
A simple hand written JSON serializer was implemented to serialize an
Aggregator and it's corresponding components and metrics. There is no
deserializer as deserialization is expected to be done in other
languages as part of the metric server clients. Since serialization is
so simple here, we decided we didn't need to pull in any dependencies.