Add a Udp based metrics server
The server responds with aggregated metrics in JSON format. A test
runs a server with no metrics and ensures it responds correctly for
correct requests, and with an error response for incorrect requests.

This server is not integrated into concord-bft yet. The aggregator that
is part of the server will be used as the aggregator for the concord-bft
library. A future commit will integrate the server as an optional
feature.