Add support for the Concord Message Format (CMF)
To enhance building message driven systems and enable architecture and
language independence, a uniform messaging format is required. This
commit defines a formal grammar for a schema language with which to
define messages. It also defines a canonical serialization format,
described informally in `messages/README.md`. Furthermore, code
generations is provided for C++ such that CMF messages can be compiled
to generate C++ structs and their corresponding serialization code.

The README provides much more information. Please read it to get a more
detailed understanding.

It's expected that this will lead to much discussion. However, please
try to focus on the semantics of the messages, and the code generation,
and don't nitpick the syntax of the messages. The schema for the
messages is formally defined by the grammar, and we can add syntax or
change naming as needed by modifying the grammar later.