Enhance logging of sent and received messages
A new logger, `MSGS`, was introduced so that debug messages for messages
sent and received by ReplicaImp and ReplicaBase could be turned on
independently. Both of those files were modified to use the new logger
instead of the global logger, `GL`.

Additionally, the name of the MsgCode is now printed instead of the
number. This makes following the logs much easier.