TLS: Allow short reads
We allow short reads of the msg size header so that we can start a read
timer and close the connection if a client only sends partial data. Note
taht we already had a read timer in place for msg reads themselves. We
started that timer after reading the msg size in the prior version,
since we don't want to force clients to send messages unncessarily with
arbitrary timers.

This code also fixes a long-standing bug by retrying message reads on 0
byte short reads without an error. This is a violation of ASIO contract
for async_read which guarantees an error or a full buffer will be
returned. This bug is fixed in later versions, and while we plan to
upgrade, we work around it in until then.

We also prevent unaligned reads of the message size header by no longer
casting out of a raw data buffer. While this is not a bug on x86
architectures, and likely doesn't even cause perf issues, it is
undefined behavior in the C++ standard, and may cause both bugs and perf
issues on other architectures. Furthermore, C++ compilers may choose to
vectorize this code which would indeed cause a real bug:
http://pzemtsov.github.io/2016/11/06/bug-story-alignment-on-x86.html