TLS: Don't run async_write outside asio thread.
async_write is a composed operation, and may result in multiple
async_write_some calls executing as a result of it. Additionally, for
TLS it is concurrently accessing the openssl context with any concurrent
async_read calls. Both of these scenarios are not thread safe and
therefore we must only run async_write in the asio thread via the use of
io_service::post.

Additionally, we cleanup lock usage and no longer hold it for the
duration of `write`. To make `queued_size_in_bytes` thread safe for
update we move it to the send method where the appropriate lock is held.
We also move some other metrics around for correctness.