Add an FSM dispatcher for value type based states
fsm.h allows the definition of state machines by providing each state
and message as a value type, and providing an explicit `Transition`
struct that defines the next state to transition to along with any
outgoing messages that result from the state transition.

An explicit decision was made to *separate IO from business logic* in
FSMs via explicit output messages during state transition. This
decision creates the following benefits:
 * States can be individually unit tested without requiring mocks
 for commmunication.
 * Messages and state transitions can be logged by a high level
 dispatcher at the point of each state transition rather than littering
 the FSM states with log messages which can obscure business logic.
 * Implementation of message transport can be changed without having to
 touch the FSM states themselves. There is no requirement to provide an
 "ITransport" interface. In fact, messages can be retrieved and "sent"
 directly via function call rather than any cross thread or network
 transport.
 * Deterministic tests can be run where all replica FSMs are run in a
 single thread and all messages traced. Deterministic adversaries can be
 introduced to alter these traces by dropping/reordering/mutating
 messages.
 * Visualization tools and interactive application level debuggers can
 be built to "step" through the protocol states and messages. Such
 visualization and debugging can be made "reversible" by caching
 intermediate states and replaying partial traces.
 * Mutation of persistent state can be modelled as messages for a disk
 manager, thus allowing different implementations of storage, and a
 concrete knowledge of exactly what data is persisted and when according
 to traces.

For this implementation in particular, the `Dispatcher` is the key FSM
type. It is repsonsible for wrapping concrete messages and states into
global `std::variant` types such that they can be *unified* into a
single type, which can be stored and acted upon in a uniform manner.

In this implementation a state is just a struct or class that has
an overloaded `on` method that takes an r-value reference to a message
and returns a `Transition`. A `Transition` is the next state, or the
next state followed by any number of output messages. Each state is
mutated by r-value reference such that a "new version" of the same state
is returned or a new state is returned. This is a "functional" style of
design that allows the use of value types exclusively and makes every
state transition, including those transitions to the same state explicit
and "externally visible".

One key benefit of this dispatcher style for FSM authors is that states
only need to know about messages and states that they transition into.
They don't need to worry about handling messages that are invalid in the
given state. Such handlers are automatically implemented during
compilation by template instantiation/SFINAE, and allow keeping state
logic minimal.

Although there is a lot of template magic inside fsm.h to make
everything work, code written outside of fsm.h does not have to
worry about templates at all. One other thing to note is that messages
are always put into an output vector that only requires a single up
front allocation if it is properly sized.

Example usage is provided in the `dispatch_success`, `dispatch_failure`,
and `dispatch_to_final_state` tests.